refactor(overlay): extract useClickOutside hook, apply to SnapPanel and WidgetSettingsPopup
